create ingredient picker component
